+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
+-   All builtin commands use new Argument attribute.
 -   List command alias renamed to 'lc'.
 
 ### Removed